put button in right spot
