+++ src/components/EditorPane/CodeEditor/component.tsx
@@ -12,21 +12,21 @@ import {
 } from "codemirror-languageserver";
 import { once } from "../../../store/codeEditorStore";
 import { WebSocketTransport } from "@open-rpc/client-js";
+import { useRuntimeDiagnostics } from "../../../hooks/useRuntimeDiagnostics";
 
 
 export const CodeEditor = () => {
   const [lspConnectionState, setLspConnectionState] = useState<any>(null);
 
   const setCurrentThreadId = usePatchStore((state) => state.setCodeThreadId);
-  const setTransport = usePatchStore((state)=>state.setTransportRef)
-
+  const setTransport = usePatchStore((state)=>state.setTransportRef);
   useEffect(() => {
-    // const serverUri = `ws://localhost:8081` as
-    //   | `ws://${string}`
-    //   | `wss://${string}`;
-    const serverUri = `${process.env.LSP_SERVER_URL}` as
+    const serverUri = `ws://localhost:8080` as
       | `ws://${string}`
       | `wss://${string}`;
+    // const serverUri = `${process.env.LSP_SERVER_URL}` as
+    //   | `ws://${string}`
+    //   | `wss://${string}`;
 
     const transport = new WebSocketTransport(serverUri);
     setTransport(transport)
